Update README.md
Belicima piatti is a mobile application that provides us with the ability to create your own masterpiece dishes, It depends on how strongly you work on It. If you're interested in creating some sort of delicious dishes and share it's receipy with each other, then you will be also interested in this app. Someone's dream will come true - the dream to grow in mastery of cooking excellent dishes. We all wait for you!